requirements.txt
flask
gunicorn==19.3.0
requests
slackclient==2.8.0
oauth2client
gspread
beautifulsoup4
lxml==4.5.2
pytest==6.0.1